refactor: rename WindowManager and WindowRegistration to snake_case
