+++ src/gui/window_manager.h
